wip: fix dupe
